AppInsightsExporterApp/requirements.txt
requests==2.23.0
pyyaml==5.3.1
prometheus_client==0.8.0
pyfiglet==0.8.post1
